Some more power handling features
It's up to user preference, but I'd rather the machine run if it's plugged in
and suspend if otherwise. Also if plugged in, turn of wifi power save.

I've had hit and miss issues with acpi not detecting lid closes properly, so, if
an event is lost, just run a quick check every minute.